Refactor: Rename modules to vetCenterHealthServices
