+++ src/data/queries/index.ts
@@ -19,7 +19,7 @@ import * as ExpandableText from '@/data/queries/expandableText'
 import * as FeaturedContent from '@/data/queries/featuredContent'
 import * as HeaderFooter from '@/data/queries/headerFooter'
 import * as HealthCareLocalFacility from '@/data/queries/healthCareLocalFacility'
-import * as HealthServices from '@/data/queries/healthServices'
+import * as HealthServices from '@/data/queries/vetCenterHealthServices'
 import * as LinkTeaser from '@/data/queries/linkTeaser'
 import * as MediaDocument from '@/data/queries/mediaDocument'
 import * as MediaImage from '@/data/queries/mediaImage'
